added the components folder and moved the componentes from the page file
